bunch more shit changed
