+++ actions/canvas_image_options_MOD.ts
@@ -59,7 +59,7 @@ const action: Action<'storage' | 'varName' | 'mirror' | 'rotation' | 'width' | '
     glob.refreshVariableList(document.getElementById('storage'));
   },
 
-  action(cache) {
+  action(this, cache) {
     const data = cache.actions[cache.index];
     const storage = parseInt(data.storage, 10);
     const varName = this.evalMessage(data.varName, cache);
