+++ actions/check_if_member_has_role_MOD.ts
@@ -1,6 +1,9 @@
-module.exports = {
-  name: 'Check Global Data',
-  section: 'Data',
+import type { Action } from '../typings/globals';
+
+const action: Action<'member' | 'varName' | 'role' | 'varName2' | 'iftrue' | 'iftrueVal' | 'iffalse' | 'iffalseVal'> = {
+  name: 'Check If Member has Role',
+  section: 'Conditions',
+  fields: ['member', 'varName', 'role', 'varName2', 'iftrue', 'iftrueVal', 'iffalse', 'iffalseVal'],
 
   subtitle(data) {
     const results = [
@@ -13,38 +16,39 @@ module.exports = {
     return `If True: ${results[parseInt(data.iftrue, 10)]} ~ If False: ${results[parseInt(data.iffalse, 10)]}`;
   },
 
-  fields: ['dataName', 'comparison', 'value', 'iftrue', 'iftrueVal', 'iffalse', 'iffalseVal', 'Jump to Anchor'],
-
-  html(_isEvent, data) {
+  html(isEvent, data) {
     return `
-<div style="padding-top: 8px;">
-  <div style="float: left; width: 50%;">
-    Data Name:<br>
-    <input id="dataName" class="round" type="text">
-  </div>
-  <div style="float: left; width: 45%;">
-    Comparison Type:<br>
-    <select id="comparison" class="round">
-      <option value="0">Exists</option>
-      <option value="1" selected>Equals</option>
-      <option value="2">Equals Exactly</option>
-      <option value="3">Less Than</option>
-      <option value="4">Greater Than</option>
-      <option value="5">Includes</option>
-      <option value="6">Matches Regex</option>
+<div>
+  <div style="float: left; width: 35%;">
+    Source Member:<br>
+    <select id="member" class="round" onchange="glob.memberChange(this, 'varNameContainer')">
+      ${data.members[isEvent ? 1 : 0]}
     </select>
   </div>
-</div><br><br><br>
-<div style="padding-top: 8px;">
-  Value to Compare to:<br>
-  <input id="value" class="round" type="text" name="is-eval">
-</div>
-<div style="padding-top: 16px;">
-  ${data.conditions[0]}
+  <div id="varNameContainer" style="display: none; float: right; width: 60%;">
+    Variable Name:<br>
+    <input id="varName" class="round" type="text" list="variableList"><br>
+    </div>
+  </div><br><br><br>
+  <div style="padding-top: 8px;">
+    <div style="float: left; width: 35%;">
+      Source Role:<br>
+      <select id="role" class="round" name="second-list" onchange="glob.roleChange(this, 'varNameContainer2')">
+        ${data.roles[isEvent ? 1 : 0]}
+      </select>
+    </div>
+    <div id="varNameContainer2" style="display: none; float: right; width: 60%;">
+      Variable Name:<br>
+      <input id="varName2" class="round" type="text" list="variableList2"><br>
+    </div>
+  </div><br><br><br>
+  <div style="padding-top: 8px;">
+    ${data.conditions[0]}
+  </div>
 </div>`;
   },
 
-  init() {
+  init(this: any) {
     const { glob, document } = this;
     const option = document.createElement('OPTION');
     option.value = '4';
@@ -58,7 +62,7 @@ module.exports = {
     const iftrue = document.getElementById('iftrue');
     if (iftrue.length === 4) iftrue.add(option2);
 
-    glob.onChangeTrue = function onChangeTrue(event) {
+    glob.onChangeTrue = function onChangeTrue(event: any) {
       switch (parseInt(event.value, 10)) {
         case 0:
         case 1:
@@ -80,7 +84,7 @@ module.exports = {
           break;
       }
     };
-    glob.onChangeFalse = function onChangeFalse(event) {
+    glob.onChangeFalse = function onChangeFalse(event: any) {
       switch (parseInt(event.value, 10)) {
         case 0:
         case 1:
@@ -102,51 +106,37 @@ module.exports = {
           break;
       }
     };
+    glob.memberChange(document.getElementById('member'), 'varNameContainer');
+    glob.roleChange(document.getElementById('role'), 'varNameContainer2');
     glob.onChangeTrue(document.getElementById('iftrue'));
     glob.onChangeFalse(document.getElementById('iffalse'));
   },
 
-  action(cache) {
+  action(this, cache) {
     const data = cache.actions[cache.index];
-    const dataName = this.evalMessage(data.dataName, cache);
-    const compare = parseInt(data.comparison, 10);
-    const { Globals } = this.getDBM();
-    const val1 = Globals.data(dataName);
-    let val2 = this.evalMessage(data.value, cache);
-    if (compare !== 6) val2 = this.eval(val2, cache);
-    if (val2 === false) val2 = this.evalMessage(data.value, cache);
+
+    const type = parseInt(data.member, 10);
+    const varName = this.evalMessage(data.varName, cache);
+    const member = this.getMember(type, varName, cache);
+
+    const type2 = parseInt(data.role, 10);
+    const varName2 = this.evalMessage(data.varName2, cache);
+    const role = this.getRole(type2, varName2, cache);
 
     let result = false;
-    switch (compare) {
-      case 0:
-        result = val1 !== undefined;
-        break;
-      case 1:
-        // eslint-disable-next-line eqeqeq
-        result = val1 == val2;
-        break;
-      case 2:
-        result = val1 === val2;
-        break;
-      case 3:
-        result = val1 < val2;
-        break;
-      case 4:
-        result = val1 > val2;
-        break;
-      case 5:
-        if (typeof val1.includes === 'function') {
-          result = val1.includes(val2);
-        }
-        break;
-      case 6:
-        result = Boolean(val1.match(new RegExp(`^${val2}$`, 'i')));
-        break;
-      default:
-        break;
+    if (role) {
+      if (Array.isArray(member)) {
+        result = member.every((mem) => {
+          return this.dest(mem, 'roles', 'cache') && mem.roles.cache.has(role.id);
+        });
+      } else if (this.dest(member, 'roles', 'cache')) {
+        result = member.roles.cache.has(role.id);
+      }
     }
     this.executeResults(result, data, cache);
   },
 
   mod() {},
 };
+
+module.exports = action;
