+++ actions/check_if_user_reacted_MOD.ts
@@ -1,6 +1,10 @@
-module.exports = {
-  name: 'Check Member Data',
-  section: 'Data',
+import type { Action } from '../typings/globals';
+
+const action: Action<
+  'member' | 'varName' | 'reaction' | 'varName2' | 'iftrue' | 'iftrueVal' | 'iffalse' | 'iffalseVal'
+> = {
+  name: 'Check If User Reacted',
+  section: 'Conditions',
 
   subtitle(data) {
     const results = [
@@ -13,57 +17,42 @@ module.exports = {
     return `If True: ${results[parseInt(data.iftrue, 10)]} ~ If False: ${results[parseInt(data.iffalse, 10)]}`;
   },
 
-  fields: ['member', 'varName', 'dataName', 'comparison', 'value', 'iftrue', 'iftrueVal', 'iffalse', 'iffalseVal'],
+  fields: ['member', 'varName', 'reaction', 'varName2', 'iftrue', 'iftrueVal', 'iffalse', 'iffalseVal'],
 
   html(isEvent, data) {
     return `
-<div><p>This action has been modified by DBM Mods.</p></div><br>
 <div>
   <div style="float: left; width: 35%;">
-    Member:<br>
+    Source Member:<br>
     <select id="member" class="round" onchange="glob.memberChange(this, 'varNameContainer')">
       ${data.members[isEvent ? 1 : 0]}
     </select>
   </div>
   <div id="varNameContainer" style="display: none; float: right; width: 60%;">
     Variable Name:<br>
-    <input id="varName" class="round" type="text" list="variableList">
+    <input id="varName" class="round" type="text" list="variableList"><br>
   </div>
-</div><br><br><br>
-<div style="padding-top: 8px;">
-  <div style="float: left; width: 50%;">
-    Data Name:<br>
-    <input id="dataName" class="round" type="text">
-  </div>
-  <div style="float: left; width: 45%;">
-    Comparison Type:<br>
-    <select id="comparison" class="round">
-      <option value="0">Exists</option>
-      <option value="1" selected>Equals</option>
-      <option value="2">Equals Exactly</option>
-      <option value="3">Less Than</option>
-      <option value="4">Greater Than</option>
-      <option value="5">Includes</option>
-      <option value="6">Matches Regex</option>
-      <option value="7">Length is Bigger Than</option>
-      <option value="8">Length is Smaller Than</option>
-      <option value="9">Length Equals</option>
-      <option value="10">Starts With</option>
-      <option value="11">Ends With</option>
+</div><br><br><br><br>
+<div>
+  <div style="float: left; width: 35%;">
+    Source Reaction:<br>
+    <select id="reaction" class="round" onchange="glob.refreshVariableList(this)">
+      ${data.variables[1]}
     </select>
   </div>
+  <div id="varNameContainer2" style="float: right; width: 60%;">
+    Variable Name:<br>
+    <input id="varName2" class="round" type="text" list="variableList"><br>
+  </div>
 </div><br><br><br>
 <div style="padding-top: 8px;">
-  Value to Compare to:<br>
-  <input id="value" class="round" type="text" name="is-eval">
-</div>
-<div style="padding-top: 16px;">
   ${data.conditions[0]}
 </div>`;
   },
 
-  init() {
+  init(this: any) {
     const { glob, document } = this;
+
     const option = document.createElement('OPTION');
     option.value = '4';
     option.text = 'Jump to Anchor';
@@ -76,7 +65,7 @@ module.exports = {
     const iftrue = document.getElementById('iftrue');
     if (iftrue.length === 4) iftrue.add(option2);
 
-    glob.onChangeTrue = function onChangeTrue(event) {
+    glob.onChangeTrue = function onChangeTrue(event: any) {
       switch (parseInt(event.value, 10)) {
         case 0:
         case 1:
@@ -98,7 +87,7 @@ module.exports = {
           break;
       }
     };
-    glob.onChangeFalse = function onChangeFalse(event) {
+    glob.onChangeFalse = function onChangeFalse(event: any) {
       switch (parseInt(event.value, 10)) {
         case 0:
         case 1:
@@ -121,71 +110,34 @@ module.exports = {
       }
     };
     glob.memberChange(document.getElementById('member'), 'varNameContainer');
+    glob.refreshVariableList(document.getElementById('reaction'));
     glob.onChangeTrue(document.getElementById('iftrue'));
     glob.onChangeFalse(document.getElementById('iffalse'));
   },
 
-  action(cache) {
+  action(this, cache) {
     const data = cache.actions[cache.index];
+
     const type = parseInt(data.member, 10);
     const varName = this.evalMessage(data.varName, cache);
     const member = this.getMember(type, varName, cache);
-    let result = false;
 
-    if (member && member.data) {
-      const dataName = this.evalMessage(data.dataName, cache);
-      const val1 = member.data(dataName);
-      const compare = parseInt(data.comparison, 10);
-      let val2 = this.evalMessage(data.value, cache);
-      if (compare !== 6) val2 = this.eval(val2, cache);
-      if (val2 === false) val2 = this.evalMessage(data.value, cache);
+    const type2 = parseInt(data.reaction, 10);
+    const varName2 = this.evalMessage(data.varName2, cache);
+    const reaction = this.getMods().getReaction(type2, varName2, cache);
 
-      switch (compare) {
-        case 0:
-          result = val1 !== undefined;
-          break;
-        case 1:
-          // eslint-disable-next-line eqeqeq
-          result = val1 == val2;
-          break;
-        case 2:
-          result = val1 === val2;
-          break;
-        case 3:
-          result = val1 < val2;
-          break;
-        case 4:
-          result = val1 > val2;
-          break;
-        case 5:
-          if (typeof val1.includes === 'function') {
-            result = val1.includes(val2);
-          }
-          break;
-        case 6:
-          result = Boolean(val1.match(new RegExp(`^${val2}$`, 'i')));
-          break;
-        case 7:
-          result = val1.length > val2;
-          break;
-        case 8:
-          result = val1.length < val2;
-          break;
-        case 9:
-          result = val1.length === val2;
-          break;
-        case 10:
-          result = val1.startsWith(val2);
-          break;
-        case 11:
-          result = val1.endsWith(val2);
-          break;
-        default:
-          break;
-      }
+    let result;
+    if (Array.isArray(member)) {
+      result = member.every(
+        (user) => this.dest(reaction, 'users', 'cache') && user && reaction.users.cache.has(user.id),
+      );
+    } else if (this.dest(reaction, 'users', 'cache') && member) {
+      result = reaction.users.cache.has(member.id);
     }
     this.executeResults(result, data, cache);
   },
 
   mod() {},
 };
+
+module.exports = action;
