+++ actions/check_parameters.ts
@@ -1,6 +1,9 @@
-module.exports = {
+import type { Action } from '../typings/globals';
+
+const action: Action<'condition' | 'comparison' | 'value' | 'iftrue' | 'iftrueVal' | 'iffalse' | 'iffalseVal'> = {
   name: 'Check Parameters',
   section: 'Conditions',
+  fields: ['condition', 'comparison', 'value', 'iftrue', 'iftrueVal', 'iffalse', 'iffalseVal'],
 
   subtitle(data) {
     const results = [
@@ -12,8 +15,6 @@ module.exports = {
     ];
     return `If True: ${results[parseInt(data.iftrue, 10)]} ~ If False: ${results[parseInt(data.iffalse, 10)]}`;
   },
-
-  fields: ['condition', 'comparison', 'value', 'iftrue', 'iftrueVal', 'iffalse', 'iffalseVal'],
 
   html(_isEvent, data) {
     return `
@@ -48,7 +49,7 @@ module.exports = {
 </div>`;
   },
 
-  init() {
+  init(this: any) {
     const { glob, document } = this;
     const option = document.createElement('OPTION');
     option.value = '4';
@@ -62,7 +63,7 @@ module.exports = {
     const iftrue = document.getElementById('iftrue');
     if (iftrue.length === 4) iftrue.add(option2);
 
-    glob.onChangeTrue = function onChangeTrue(event) {
+    glob.onChangeTrue = function onChangeTrue(event: any) {
       switch (parseInt(event.value, 10)) {
         case 0:
         case 1:
@@ -84,7 +85,7 @@ module.exports = {
           break;
       }
     };
-    glob.onChangeFalse = function onChangeFalse(event) {
+    glob.onChangeFalse = function onChangeFalse(event: any) {
       switch (parseInt(event.value, 10)) {
         case 0:
         case 1:
@@ -110,12 +111,12 @@ module.exports = {
     glob.onChangeFalse(document.getElementById('iffalse'));
   },
 
-  action(cache) {
+  action(this, cache) {
     const data = cache.actions[cache.index];
     const { msg } = cache;
     let result = false;
 
-    if (msg && msg.content.length > 0) {
+    if (msg.content.length > 0) {
       const condition = parseInt(data.condition, 10);
       let value = 0;
       switch (condition) {
@@ -123,7 +124,7 @@ module.exports = {
           value = msg.content.split(/\s+/).length - 1;
           break;
         case 1:
-          value = msg.mentions.members.array().length;
+          value = msg.mentions.members?.array().length ?? 0;
           break;
         case 2:
           value = msg.mentions.channels.array().length;
@@ -163,3 +164,5 @@ module.exports = {
 
   mod() {},
 };
+
+module.exports = action;
