+++ package.json
@@ -14,6 +14,7 @@
     "docs:watch": "typedoc --watch --tsconfig tsconfig.json"
   },
   "devDependencies": {
+    "@types/jsonpath": "^0.2.0",
     "@types/node": "^16.3.1",
     "@types/opentype.js": "^1.3.2",
     "@typescript-eslint/eslint-plugin": "^4.28.4",
@@ -36,6 +37,7 @@
   "dependencies": {
     "canvas": "^2.8.0",
     "imagedata-filters": "^1.0.1",
+    "jsonpath": "^1.1.1",
     "opentype.js": "^1.3.3",
     "typedoc": "^0.21.4",
     "typedoc-darker-theme": "^1.2.1"
